backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
sqlalchemy[asyncio]
psycopg2-binary==2.9.9
pydantic==2.5.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
alembic==1.12.1
python-dotenv==1.0.0
email-validator==2.1.0.post1
asyncpg

# LLM Dependencies
openai>=1.6.1 